+++ apollo-common/src/test/java/com/ctrip/framework/apollo/common/exception/NotFoundExceptionTest.java
@@ -16,8 +16,9 @@
  */
 package com.ctrip.framework.apollo.common.exception;
 
-import org.junit.Assert;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 
 public class NotFoundExceptionTest {
 
@@ -33,49 +34,54 @@ public class NotFoundExceptionTest {
         clusterName, namespaceName, key);
     e2 = new NotFoundException(
         String.format("item not found for %s %s %s %s", appId, clusterName, namespaceName, key));
-    Assert.assertEquals(e1.getMessage(), e2.getMessage());
+    assertEquals(e1.getMessage(), e2.getMessage());
   }
 
   @Test
   public void testAppNotFoundException() {
     NotFoundException exception = NotFoundException.appNotFound(appId);
-    Assert.assertEquals(exception.getMessage(), "app not found for appId:app-1001");
+    assertEquals(exception.getMessage(), "app not found for appId:app-1001");
   }
 
   @Test
   public void testClusterNotFoundException() {
     NotFoundException exception = NotFoundException.clusterNotFound(appId, clusterName);
-    Assert.assertEquals(exception.getMessage(), "cluster not found for appId:app-1001 clusterName:test");
+    assertEquals(exception.getMessage(), "cluster not found for appId:app-1001 clusterName:test");
   }
 
   @Test
   public void testNamespaceNotFoundException() {
     NotFoundException exception = NotFoundException.namespaceNotFound(appId, clusterName, namespaceName);
-    Assert.assertEquals(exception.getMessage(), "namespace not found for appId:app-1001 clusterName:test namespaceName:application");
+    assertEquals(exception.getMessage(), "namespace not found for appId:app-1001 clusterName:test namespaceName:application");
 
     exception = NotFoundException.namespaceNotFound(66);
-    Assert.assertEquals(exception.getMessage(), "namespace not found for namespaceId:66");
+    assertEquals(exception.getMessage(), "namespace not found for namespaceId:66");
   }
 
   @Test
   public void testReleaseNotFoundException() {
     NotFoundException exception = NotFoundException.releaseNotFound(66);
-    Assert.assertEquals(exception.getMessage(), "release not found for releaseId:66");
+    assertEquals(exception.getMessage(), "release not found for releaseId:66");
   }
 
   @Test
   public void testItemNotFoundException(){
     NotFoundException exception = NotFoundException.itemNotFound(66);
-    Assert.assertEquals(exception.getMessage(), "item not found for itemId:66");
+    assertEquals(exception.getMessage(), "item not found for itemId:66");
 
     exception = NotFoundException.itemNotFound("test.key");
-    Assert.assertEquals(exception.getMessage(), "item not found for itemKey:test.key");
+    assertEquals(exception.getMessage(), "item not found for itemKey:test.key");
 
     exception = NotFoundException.itemNotFound(appId, clusterName, namespaceName, "test.key");
-    Assert.assertEquals(exception.getMessage(), "item not found for appId:app-1001 clusterName:test namespaceName:application itemKey:test.key");
+    assertEquals(exception.getMessage(), "item not found for appId:app-1001 clusterName:test namespaceName:application itemKey:test.key");
 
     exception = NotFoundException.itemNotFound(appId, clusterName, namespaceName, 66);
-    Assert.assertEquals(exception.getMessage(), "item not found for appId:app-1001 clusterName:test namespaceName:application itemId:66");
+    assertEquals(exception.getMessage(), "item not found for appId:app-1001 clusterName:test namespaceName:application itemId:66");
   }
 
+  @Test
+  void roleNotFound() {
+    NotFoundException exception = NotFoundException.roleNotFound("CreateApplication+SystemRole");
+    assertEquals(exception.getMessage(), "role not found for roleName:CreateApplication+SystemRole, please check apollo portal DB table 'Role'");
+  }
 }