+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/openapi/auth/ConsumerPermissionValidator.java
@@ -16,6 +16,8 @@
  */
 package com.ctrip.framework.apollo.openapi.auth;
 
+import static com.ctrip.framework.apollo.portal.service.SystemRoleManagerService.SYSTEM_PERMISSION_TARGET_ID;
+
 import com.ctrip.framework.apollo.openapi.service.ConsumerRolePermissionService;
 import com.ctrip.framework.apollo.openapi.util.ConsumerAuthUtil;
 import com.ctrip.framework.apollo.portal.constant.PermissionType;
@@ -69,5 +71,10 @@ public class ConsumerPermissionValidator {
   public boolean hasCreateClusterPermission(HttpServletRequest request, String appId) {
     return permissionService.consumerHasPermission(consumerAuthUtil.retrieveConsumerId(request),
         PermissionType.CREATE_CLUSTER, appId);
+  }
+
+  public boolean hasCreateApplicationPermission(HttpServletRequest request) {
+    long consumerId = consumerAuthUtil.retrieveConsumerId(request);
+    return permissionService.consumerHasPermission(consumerId, PermissionType.CREATE_APPLICATION, SYSTEM_PERMISSION_TARGET_ID);
   }
 }