+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/openapi/v1/controller/AppController.java
@@ -16,14 +16,19 @@
  */
 package com.ctrip.framework.apollo.openapi.v1.controller;
 
+import com.ctrip.framework.apollo.common.exception.BadRequestException;
 import com.ctrip.framework.apollo.openapi.api.AppOpenApiService;
+import com.ctrip.framework.apollo.openapi.dto.OpenCreateAppDTO;
 import com.ctrip.framework.apollo.openapi.service.ConsumerService;
 import com.ctrip.framework.apollo.openapi.util.ConsumerAuthUtil;
 import com.ctrip.framework.apollo.openapi.dto.OpenAppDTO;
 import com.ctrip.framework.apollo.openapi.dto.OpenEnvClusterDTO;
+import com.ctrip.framework.apollo.portal.entity.model.AppModel;
 import java.util.Arrays;
 import java.util.Set;
 import javax.servlet.http.HttpServletRequest;
+import javax.transaction.Transactional;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
@@ -45,6 +50,31 @@ public class AppController {
     this.consumerAuthUtil = consumerAuthUtil;
     this.consumerService = consumerService;
     this.appOpenApiService = appOpenApiService;
+  }
+
+  /**
+   * @see com.ctrip.framework.apollo.portal.controller.AppController#create(AppModel)
+   */
+  @Transactional
+  @PreAuthorize(value = "@consumerPermissionValidator.hasCreateApplicationPermission(#request)")
+  @PostMapping(value = "/apps")
+  public void createApp(
+      @RequestBody OpenCreateAppDTO req,
+      HttpServletRequest request
+  ) {
+    if (null == req.getApp()) {
+      throw new BadRequestException("App is null");
+    }
+    final OpenAppDTO app = req.getApp();
+    if (null == app.getAppId()) {
+      throw new BadRequestException("AppId is null");
+    }
+    // create app
+    this.appOpenApiService.createApp(req);
+    if (req.isAssignAppRoleToSelf()) {
+      long consumerId = this.consumerAuthUtil.retrieveConsumerId(request);
+      consumerService.assignAppRoleToConsumer(consumerId, app.getAppId());
+    }
   }
 
   @GetMapping(value = "/apps/{appId}/envclusters")