+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/controller/ConsumerController.java
@@ -23,6 +23,8 @@ import com.ctrip.framework.apollo.openapi.entity.Consumer;
 import com.ctrip.framework.apollo.openapi.entity.ConsumerRole;
 import com.ctrip.framework.apollo.openapi.entity.ConsumerToken;
 import com.ctrip.framework.apollo.openapi.service.ConsumerService;
+import com.ctrip.framework.apollo.portal.entity.vo.consumer.ConsumerCreateRequestVO;
+import com.ctrip.framework.apollo.portal.entity.vo.consumer.ConsumerInfo;
 import com.ctrip.framework.apollo.portal.environment.Env;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
@@ -48,27 +50,48 @@ public class ConsumerController {
     this.consumerService = consumerService;
   }
 
+  private Consumer convertToConsumer(ConsumerCreateRequestVO requestVO) {
+    Consumer consumer = new Consumer();
+    consumer.setAppId(requestVO.getAppId());
+    consumer.setName(requestVO.getName());
+    consumer.setOwnerName(requestVO.getOwnerName());
+    consumer.setOrgId(requestVO.getOrgId());
+    consumer.setOrgName(requestVO.getOrgName());
+    return consumer;
+  }
 
   @Transactional
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
   @PostMapping(value = "/consumers")
-  public ConsumerToken createConsumer(@RequestBody Consumer consumer,
-                                      @RequestParam(value = "expires", required = false)
-                                      @DateTimeFormat(pattern = "yyyyMMddHHmmss") Date
-                                          expires) {
-
-    if (StringUtils.isContainEmpty(consumer.getAppId(), consumer.getName(),
-                                   consumer.getOwnerName(), consumer.getOrgId())) {
-      throw new BadRequestException("Params(appId、name、ownerName、orgId) can not be empty.");
+  public ConsumerInfo create(
+      @RequestBody ConsumerCreateRequestVO requestVO,
+      @RequestParam(value = "expires", required = false)
+      @DateTimeFormat(pattern = "yyyyMMddHHmmss") Date expires
+  ) {
+    if (StringUtils.isBlank(requestVO.getAppId())) {
+      throw BadRequestException.appIdIsBlank();
+    }
+    if (StringUtils.isBlank(requestVO.getName())) {
+      throw BadRequestException.appNameIsBlank();
+    }
+    if (StringUtils.isBlank(requestVO.getOwnerName())) {
+      throw BadRequestException.ownerNameIsBlank();
+    }
+    if (StringUtils.isBlank(requestVO.getOrgId())) {
+      throw BadRequestException.orgIdIsBlank();
     }
 
-    Consumer createdConsumer = consumerService.createConsumer(consumer);
+    Consumer createdConsumer = consumerService.createConsumer(convertToConsumer(requestVO));
 
     if (Objects.isNull(expires)) {
       expires = DEFAULT_EXPIRES;
     }
 
-    return consumerService.generateAndSaveConsumerToken(createdConsumer, expires);
+    ConsumerToken consumerToken = consumerService.generateAndSaveConsumerToken(createdConsumer, expires);
+    if (requestVO.isAllowCreateApplication()) {
+      consumerService.assignCreateApplicationRoleToConsumer(consumerToken.getToken());
+    }
+    return consumerService.getConsumerInfoByAppId(requestVO.getAppId());
   }
 
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
@@ -78,11 +101,19 @@ public class ConsumerController {
   }
 
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
+  @GetMapping(value = "/consumer/info/by-appId")
+  public ConsumerInfo getConsumerInfoByAppId(@RequestParam String appId) {
+    return consumerService.getConsumerInfoByAppId(appId);
+  }
+
+  @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
   @PostMapping(value = "/consumers/{token}/assign-role")
-  public List<ConsumerRole> assignNamespaceRoleToConsumer(@PathVariable String token,
-                                                          @RequestParam String type,
-                                                          @RequestParam(required = false) String envs,
-                                                          @RequestBody NamespaceDTO namespace) {
+  public List<ConsumerRole> assignNamespaceRoleToConsumer(
+      @PathVariable String token,
+      @RequestParam String type,
+      @RequestParam(required = false) String envs,
+      @RequestBody NamespaceDTO namespace) {
+    List<ConsumerRole> consumerRoleList = new ArrayList<>(8);
 
     String appId = namespace.getAppId();
     String namespaceName = namespace.getNamespaceName();
@@ -117,13 +148,16 @@ public class ConsumerController {
       return consumeRoles;
     }
 
-    return consumerService.assignNamespaceRoleToConsumer(token, appId, namespaceName);
+    consumerRoleList.addAll(
+        consumerService.assignNamespaceRoleToConsumer(token, appId, namespaceName)
+    );
+    return consumerRoleList;
   }
 
   @GetMapping("/consumers")
   @PreAuthorize(value = "@permissionValidator.isSuperAdmin()")
-  public List<Consumer> getConsumerList(Pageable page){
-    return consumerService.findAllConsumer(page);
+  public List<ConsumerInfo> getConsumerList(Pageable page){
+    return consumerService.findConsumerInfoList(page);
   }
 
   @DeleteMapping(value = "/consumers/by-appId")