+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/AppService.java
@@ -22,16 +22,20 @@ import com.ctrip.framework.apollo.common.entity.App;
 import com.ctrip.framework.apollo.common.exception.BadRequestException;
 import com.ctrip.framework.apollo.common.utils.BeanUtils;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
+import com.ctrip.framework.apollo.portal.api.AdminServiceAPI.AppAPI;
 import com.ctrip.framework.apollo.portal.environment.Env;
 import com.ctrip.framework.apollo.portal.api.AdminServiceAPI;
 import com.ctrip.framework.apollo.portal.constant.TracerEventType;
 import com.ctrip.framework.apollo.portal.entity.bo.UserInfo;
 import com.ctrip.framework.apollo.portal.entity.vo.EnvClusterInfo;
+import com.ctrip.framework.apollo.portal.listener.AppCreationEvent;
 import com.ctrip.framework.apollo.portal.repository.AppRepository;
 import com.ctrip.framework.apollo.portal.spi.UserInfoHolder;
 import com.ctrip.framework.apollo.portal.spi.UserService;
+import com.ctrip.framework.apollo.portal.util.RoleUtils;
 import com.ctrip.framework.apollo.tracer.Tracer;
 import com.google.common.collect.Lists;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -39,6 +43,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Set;
+import org.springframework.util.CollectionUtils;
 
 @Service
 public class AppService {
@@ -53,16 +58,18 @@ public class AppService {
   private final FavoriteService favoriteService;
   private final UserService userService;
 
+  private final ApplicationEventPublisher publisher;
+
   public AppService(
       final UserInfoHolder userInfoHolder,
-      final AdminServiceAPI.AppAPI appAPI,
+      final AppAPI appAPI,
       final AppRepository appRepository,
       final ClusterService clusterService,
       final AppNamespaceService appNamespaceService,
       final RoleInitializationService roleInitializationService,
       final RolePermissionService rolePermissionService,
       final FavoriteService favoriteService,
-      final UserService userService) {
+      final UserService userService, ApplicationEventPublisher publisher) {
     this.userInfoHolder = userInfoHolder;
     this.appAPI = appAPI;
     this.appRepository = appRepository;
@@ -72,6 +79,7 @@ public class AppService {
     this.rolePermissionService = rolePermissionService;
     this.favoriteService = favoriteService;
     this.userService = userService;
+    this.publisher = publisher;
   }
 
 
@@ -124,8 +132,7 @@ public class AppService {
     appAPI.createApp(env, appDTO);
   }
 
-  @Transactional
-  public App createAppInLocal(App app) {
+  private App createAppInLocal(App app) {
     String appId = app.getAppId();
     App managedApp = appRepository.findByAppId(appId);
 
@@ -149,6 +156,23 @@ public class AppService {
     roleInitializationService.initAppRoles(createdApp);
 
     Tracer.logEvent(TracerEventType.CREATE_APP, appId);
+
+    return createdApp;
+  }
+
+  @Transactional
+  public App createAppAndAddRolePermission(
+      App app, Set<String> admins
+  ) {
+    App createdApp = this.createAppInLocal(app);
+
+    publisher.publishEvent(new AppCreationEvent(createdApp));
+
+    if (!CollectionUtils.isEmpty(admins)) {
+      rolePermissionService
+          .assignRoleToUsers(RoleUtils.buildAppMasterRoleName(createdApp.getAppId()),
+              admins, userInfoHolder.getUser().getUserId());
+    }
 
     return createdApp;
   }